Registre

+++ src/Models/CustomerRegistruRow.php
@@ -91,8 +91,11 @@ class CustomerRegistruRow extends Model {
     }
     
     public static function doDelete($input, $record) {
-        $record->values()->delete();
-        $record->delete();
+        if($record)
+        {
+            $record->values()->delete();
+            $record->delete();
+        }
         return $record;
     }
